+++ src/main/kotlin/com/grimfox/gec/filter/CoastlineFilter.kt
@@ -273,7 +273,7 @@ interface CoastlineFilter : Runnable {
             subtotal += weight
             degreeWeights.add(weight.toFloat())
         }
-        val total = subtotal.toFloat()
+        val total = subtotal
         subtotal = 0.0f
         for (j in 0..5) {
             subtotal += degreeWeights[j] / total
